upload-medical-documents-webhook/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
google-cloud-storage==2.10.0
google-generativeai==0.3.0
requests==2.31.0
pdfplumber==0.10.3
Werkzeug==3.0.1
gunicorn==21.2.0
